+++ compile.js
@@ -7,9 +7,6 @@
 const WaveFile = require("./index.js");
 
 /*
-WaveFile.prototype["container"] = WaveFile.prototype.container;
-WaveFile.prototype["chunkSize"] = WaveFile.prototype.chunkSize;
-WaveFile.prototype["format"] = WaveFile.prototype.format;
 WaveFile.prototype["fromScratch"] = WaveFile.prototype.fromScratch;
 WaveFile.prototype["fromBuffer"] = WaveFile.prototype.fromBuffer;
 WaveFile.prototype["toBuffer"] = WaveFile.prototype.toBuffer;
@@ -24,6 +21,18 @@ WaveFile.prototype["toALaw"] = WaveFile.prototype.toALaw;
 WaveFile.prototype["fromALaw"] = WaveFile.prototype.fromALaw;
 WaveFile.prototype["toMuLaw"] = WaveFile.prototype.toMuLaw;
 WaveFile.prototype["fromMuLaw"] = WaveFile.prototype.fromMuLaw;
+
+WaveFile.prototype["container"] = WaveFile.prototype.container;
+WaveFile.prototype["chunkSize"] = WaveFile.prototype.chunkSize;
+WaveFile.prototype["format"] = WaveFile.prototype.format;
+WaveFile.prototype["fmt"] = WaveFile.prototype.fmt;
+WaveFile.prototype["fact"] = WaveFile.prototype.fact;
+WaveFile.prototype["bext"] = WaveFile.prototype.bext;
+WaveFile.prototype["ds64"] = WaveFile.prototype.ds64;
+WaveFile.prototype["cue"] = WaveFile.prototype.cue;
+WaveFile.prototype["data"] = WaveFile.prototype.data;
+WaveFile.prototype["isInterleaved"] = WaveFile.prototype.isInterleaved;
+WaveFile.prototype["bitDepth"] = WaveFile.prototype.bitDepth;
 */
 
 window["WaveFile"] = WaveFile;