+++ test/codecs/alaw.js
@@ -5,6 +5,73 @@
 
 var assert = require('assert');
 
+
+describe('8-bit a-law to 24-bit (a-law encoded with WaveFile)', function() {
+    let fs = require("fs");
+    const WaveFile = require("../../test/loader.js");
+    let path = "test/files/";
+    
+    let wav = new WaveFile(
+        fs.readFileSync(path + "/8bit-alaw-8kHz-noBext-mono-encoded.wav"));
+
+    wav.fromALaw("24");
+
+    let wav2 = new WaveFile(wav.toBuffer());
+    fs.writeFileSync(path + "/out/24bit-8kHz-noBext-mono-decoded-alaw.wav",
+        wav2.toBuffer());
+
+    it("chunkId should be 'RIFF'",
+            function() {
+        assert.equal(wav2.container, "RIFF");
+    });
+    it("fmtChunkId should be 'fmt '",
+            function() {
+        assert.equal(wav2.fmt.chunkId, "fmt ");
+    });
+    it("format should be 'WAVE'",
+            function() {
+        assert.equal(wav2.format, "WAVE");
+    });
+    it("fmtChunkSize should be 16",
+            function() {
+        assert.equal(wav2.fmt.chunkSize, 16);
+    });
+    it("audioFormat should be 1 (PCM)",
+            function() {
+        assert.equal(wav2.fmt.audioFormat, 1);
+    });
+    it("numChannels should be 1",
+            function() {
+        assert.equal(wav2.fmt.numChannels, 1);
+    });
+    it("sampleRate should be 8000",
+            function() {
+        assert.equal(wav2.fmt.sampleRate, 8000);
+    });
+
+    it("blockAlign should be 3",
+            function() {
+        assert.equal(wav2.fmt.blockAlign, 3);
+    });
+    it("bitsPerSample should be 24",
+            function() {
+        assert.equal(wav2.fmt.bitsPerSample, 24);
+    });
+    
+    it("dataChunkId should be 'data'",
+            function() {
+        assert.equal(wav2.data.chunkId, 'data');
+    });
+    it("dataChunkSize should be > 0",
+            function() {
+        assert.ok(wav2.data.chunkSize > 0);
+    });
+    it("samples.length should be > 0",
+            function() {
+        assert.ok(wav2.data.samples.length > 0);
+    });
+});
+
 describe('16-bit to 8-bit a-law', function() {
     let fs = require("fs");
     const WaveFile = require("../../test/loader.js");